Move serializable into _internal
