+++ src/globus_sdk/_internal/guards.py
@@ -10,7 +10,7 @@ from globus_sdk._types import UUIDLike
 from globus_sdk.exc.base import ValidationError
 
 if t.TYPE_CHECKING:
-    from globus_sdk._serializable import Serializable
+    from globus_sdk._internal.serializable import Serializable
 
 if sys.version_info >= (3, 10):
     from typing import TypeGuard
