+++ src/globus_sdk/_internal/serializable.py
