+++ src/globus_sdk/token_storage/v2/token_data.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import typing as t
 
 from globus_sdk._internal.guards import validators
-from globus_sdk._serializable import Serializable
+from globus_sdk._internal.serializable import Serializable
 
 
 class TokenStorageData(Serializable):
