+++ tests/non-pytest/lazy-imports/test_modules_do_not_require_requests.py
@@ -38,8 +38,8 @@ PYTHON_BINARY = os.environ.get("GLOBUS_TEST_PY", sys.executable)
         "_internal.classprop",
         "_internal.guards",
         "_internal.remarshal",
+        "_internal.serializable",
         "_missing",
-        "_serializable",
         "_types",
         "_utils",
     ),
